Fixed issue where 2019 and earlier saved data would break due to the speed trap data not existing, but it was assumed it would be.
